Intelligent_Patient_Helper-main/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0 
flask==2.0.1
flask-cors==3.0.10
sqlite3==0.0.0
openai==1.12.0
httpx==0.28.0